resumeparser+scorer/resume_matcher/requirements.txt
python-dotenv
openai
spacy
scikit-learn
sentence-transformers
PyPDF2
python-docx
fastapi
uvicorn
pydantic
